requirements.txt
ipykernel==6.29.4
pillow==10.3.0
ultralytics==8.3.78
manga-ocr==0.1.14
fastapi
uvicorn==0.30.0
torch==2.6.0
